--- app/src/main/java/com/tpov/schoolquiz/MainViewModel.kt
@@ -1,24 +1,0 @@
-package com.tpov.schoolquiz
-
-import androidx.lifecycle.*
-import com.tpov.schoolquiz.data.database.CrimeDatabase
-import com.tpov.schoolquiz.data.database.entities.ApiQuestion
-import com.tpov.schoolquiz.data.database.entities.Question
-import com.tpov.schoolquiz.data.database.entities.Quiz
-import com.tpov.schoolquiz.data.database.entities.QuizDetail
-import kotlinx.coroutines.launch
-
-open class MainViewModel(database: CrimeDatabase) : ViewModel() {
-
-
-
-    class MainViewModelFactory(private val database: CrimeDatabase) : ViewModelProvider.Factory {
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
-                @Suppress("UNCHECKED.CAST")
-                return MainViewModel(database) as T
-            }
-            throw IllegalAccessException("Unknown ViewModelClass")
-        }
-    }
-}
