+++ app/src/main/java/com/tpov/schoolquiz/data/database/CrimeDao.kt
@@ -1,5 +1,6 @@
 package com.tpov.schoolquiz.data.database
 
+import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import com.tpov.schoolquiz.data.database.entities.QuizDetail
@@ -23,24 +24,24 @@ interface CrimeDao {
     suspend fun insertGenerateQuestion(name: List<ApiQuestion>)
 
     @Query("SELECT * FROM table_data")
-    fun getAllCrime() : Flow<List<QuizDetail>>     //Запускаем через Корутину
+    fun getAllCrime() : LiveData<List<QuizDetail>>    //Запускаем через Корутину
     @Query("SELECT * FROM table_data WHERE idNameQuiz LIKE :idName")
     fun getNameCrime(idName: String?) : Flow<List<QuizDetail>>?                     //Запускаем через Корутину
     @Query("SELECT * FROM front_list")
     fun getAllCrimeFrontList() : Flow<List<Quiz>>     //Запускаем через Корутину
     @Query("SELECT* FROM new_user_table")
-    fun getCrimeNewQuiz() : Flow<List<Question>>
+    fun getCrimeNewQuiz() : LiveData<List<Question>>
     @Query("SELECT * FROM table_generate_question")
     fun getAllGenerateQuestion(): Flow<List<ApiQuestion>>
 
     @Query("SELECT * FROM table_data")
     suspend fun getQuiz() : List<QuizDetail>
     @Query("SELECT * FROM new_user_table")
-    suspend fun getAllIdQuestion() : List<Question>
+    fun getAllIdQuestion() : LiveData<List<Question>>
     @Query("SELECT * FROM front_list")
-    suspend fun getFrontList() : List<Quiz>
+    fun getFrontList() : LiveData<List<Quiz>>
     @Query("SELECT * FROM table_generate_question")
-    suspend fun getGenerateQuestion(): List<ApiQuestion>
+    fun getGenerateQuestion(): List<ApiQuestion>
 
     @Query("SELECT * FROM new_user_table WHERE idListNameQuestion LIKE :idGeoQuiz")
     fun getGeoQuiz(idGeoQuiz: String): Flow<List<Question>>
@@ -70,7 +71,7 @@ interface CrimeDao {
     @Update
     suspend fun updateCrime(quizDetail: QuizDetail)
     @Update
-    suspend fun updateFrontList(quiz: Quiz)
+    fun updateFrontList(quiz: Quiz)
     @Update
     suspend fun updateGenerationQuestion(generateQuestion: ApiQuestion)
