+++ app/src/main/java/com/tpov/schoolquiz/domain/DeleteQuizUseCase.kt
@@ -1,8 +1,6 @@
 package com.tpov.schoolquiz.domain
 
 class DeleteQuizUseCase(private val repository: Repository) {
-
     suspend operator fun invoke(deleteAnswerQuestion: Boolean, nameQuiz: String) =
         repository.deleteQuiz(deleteAnswerQuestion, nameQuiz)
-
 }
