+++ app/src/main/java/com/tpov/schoolquiz/domain/GetInfoQuestionParamsUseCase.kt
@@ -1,0 +1,8 @@
+package com.tpov.schoolquiz.domain
+
+import com.tpov.schoolquiz.data.database.entities.QuizDetail
+
+class GetInfoQuestionParamsUseCase(private val repository: Repository) {
+    suspend operator fun invoke(updateQuestion: Boolean, insertQuiz: QuizDetail, idUser: String) =
+        repository.getInfoQuestionParams(updateQuestion, insertQuiz, idUser)
+}
