+++ app/src/main/java/com/tpov/schoolquiz/domain/GetQuestionDayUseCase.kt
@@ -1,0 +1,5 @@
+package com.tpov.schoolquiz.domain
+
+class GetQuestionDayUseCase(private val repository: Repository) {
+    suspend operator fun invoke() = repository.getQuestionDay()
+}
