+++ app/src/main/java/com/tpov/schoolquiz/domain/GetQuestionUseCase.kt
@@ -1,5 +1,5 @@
 package com.tpov.schoolquiz.domain
 
 class GetQuestionUseCase(private val repository: Repository) {
-    operator fun invoke(questionId: Int) = repository.getQuestion(questionId)
+    operator fun invoke() = repository.getQuestion()
 }
