+++ app/src/main/java/com/tpov/schoolquiz/domain/GetQuizUseCase.kt
@@ -1,0 +1,5 @@
+package com.tpov.schoolquiz.domain
+
+class GetQuizUseCase(private val repository: Repository) {
+    operator fun invoke() = repository.getQuiz()
+}
