+++ app/src/main/java/com/tpov/schoolquiz/domain/InsertApiQuestionUseCase.kt
@@ -1,0 +1,7 @@
+package com.tpov.schoolquiz.domain
+
+import com.tpov.schoolquiz.data.database.entities.ApiQuestion
+
+class InsertApiQuestionUseCase(private val repository: Repository) {
+    suspend operator fun invoke(apiQuestion: List<ApiQuestion>) = repository.insertQuestionDay(apiQuestion)
+}
