--- app/src/main/java/com/tpov/schoolquiz/domain/LoadQuestionDayUseCase.kt
@@ -1,5 +1,0 @@
-package com.tpov.schoolquiz.domain
-
-class LoadQuestionDayUseCase(private val repository: Repository) {
-    operator fun invoke(systemDate: String) = repository.getQuestionDay(systemDate)
-}
