--- app/src/main/java/com/tpov/schoolquiz/domain/NewQuizUseCase.kt
@@ -1,8 +1,0 @@
-package com.tpov.schoolquiz.domain
-
-import com.tpov.schoolquiz.data.database.entities.Question
-import com.tpov.schoolquiz.data.database.entities.Quiz
-
-class NewQuizUseCase(private val repository: Repository) {
-    operator fun invoke(quiz: Quiz, question: Question) = repository.newQuiz(quiz, question)
-}
