--- app/src/main/java/com/tpov/schoolquiz/domain/ResultQuizUseCase.kt
@@ -1,8 +1,0 @@
-package com.tpov.schoolquiz.domain
-
-import com.tpov.schoolquiz.data.database.entities.Quiz
-
-class ResultQuizUseCase(private val repository: Repository) {
-
-    operator fun invoke(quiz: Quiz) = repository.resultQuiz(quiz)
-}
